listener: preserve failure causes and isolate error delivery
